refactor: delete l1/ directory and core/scan.ts
Remove the entire ALPHA-blockchain layer (address/tx/Fulcrum network/vesting/
crypto) and the now-orphaned Fulcrum balance scanner. Relocate the dual-use
generateAddressFromMasterKey (legacy wif_hmac key derivation, also yields the
L3 chainPubkey) into core/crypto.ts so legacy-wallet loading still works.

Drop the L1/scan/electrum re-exports from index.ts and the L1 fields from
core/discover.ts (discovery is now transport/Nostr-binding only). Delete the
l1/ unit tests and repoint the few tests that used l1/ helpers to core/crypto.
