+++ core/index.ts
@@ -1,5 +1,4 @@
 export * from './Sphere';
-export * from './scan';
 export * from './discover';
 export * from './crypto';
 export * from './encryption';
