+++ index.ts
@@ -60,10 +60,6 @@ export type {
   InitProgressStep,
   InitProgress,
   InitProgressCallback,
-  ScanAddressProgress,
-  ScannedAddressResult,
-  ScanAddressesOptions,
-  ScanAddressesResult,
   DiscoverAddressProgress,
   DiscoveredAddress,
   DiscoverAddressesOptions,
@@ -336,9 +332,6 @@ export {
   // IPFS
   DEFAULT_IPFS_GATEWAYS,
   DEFAULT_IPFS_BOOTSTRAP_PEERS,
-  // L1 (ALPHA Blockchain)
-  DEFAULT_ELECTRUM_URL,
-  TEST_ELECTRUM_URL,
   // Wallet
   DEFAULT_DERIVATION_PATH,
   COIN_TYPES,
@@ -433,13 +426,6 @@ export type {
   SpentTokenResult,
   ValidationResult as TokenValidationResult,
 } from './validation';
-
-// =============================================================================
-// L1 SDK (ALPHA Blockchain)
-// =============================================================================
-
-// L1 Low-level SDK
-export * as L1 from './l1';
 
 // =============================================================================
 // Token Registry
@@ -554,15 +540,8 @@ export {
 } from './core/encryption';
 export type { EncryptedData } from './core/encryption';
 
-// L1 low-level crypto + address utilities (in addition to L1.* namespace export)
-export {
-  generatePrivateKey,
-  hexToWIF,
-  encryptWallet,
-  decryptWallet,
-} from './l1/crypto';
-
-export { generateAddressFromMasterKey } from './l1/address';
+// Legacy wallet derivation helper (dual-use: derives chainPubkey in wif_hmac mode)
+export { generateAddressFromMasterKey } from './core/crypto';
 
 // Accounting module types
 export type {
