+++ tests/integration/history-sync.test.ts
@@ -30,13 +30,6 @@ import type { ProviderStatus } from '../../types';
 import type { SaveResult, LoadResult, SyncResult } from '../../storage';
 import { TEST_NETWORK } from '../test-network';
 
-// Mock L1 to avoid real WebSocket connections
-vi.mock('../../l1/network', () => ({
-  connect: vi.fn().mockResolvedValue(undefined),
-  disconnect: vi.fn(),
-  isWebSocketConnected: vi.fn().mockReturnValue(false),
-}));
-
 // =============================================================================
 // Test directories
 // =============================================================================
