--- tests/unit/l1/addressToScriptHash.test.ts
@@ -1,96 +1,0 @@
-/**
- * Tests for l1/addressToScriptHash.ts
- * Covers Electrum scripthash conversion
- */
-
-import { describe, it, expect } from 'vitest';
-import { addressToScriptHash } from '../../../l1/addressToScriptHash';
-import { encodeBech32 } from '../../../core/bech32';
-
-// =============================================================================
-// addressToScriptHash Tests
-// =============================================================================
-
-describe('addressToScriptHash()', () => {
-  it('should convert bech32 address to Electrum scripthash', () => {
-    // Create a known address
-    const program = new Uint8Array([
-      0x75, 0x1e, 0x76, 0xe8, 0x19, 0x91, 0x96, 0xd4, 0x54, 0x94, 0x1c, 0x45, 0xd1, 0xb3, 0xa3,
-      0x23, 0xf1, 0x43, 0x3b, 0xd6,
-    ]);
-    const address = encodeBech32('alpha', 0, program);
-
-    const scripthash = addressToScriptHash(address);
-
-    // Should be 64 hex chars (32 bytes reversed)
-    expect(scripthash).toHaveLength(64);
-    expect(/^[0-9a-f]+$/.test(scripthash)).toBe(true);
-  });
-
-  it('should produce deterministic result', () => {
-    const program = new Uint8Array(20).fill(0xab);
-    const address = encodeBech32('alpha', 0, program);
-
-    const hash1 = addressToScriptHash(address);
-    const hash2 = addressToScriptHash(address);
-
-    expect(hash1).toBe(hash2);
-  });
-
-  it('should produce different scripthash for different addresses', () => {
-    const program1 = new Uint8Array(20).fill(0xab);
-    const program2 = new Uint8Array(20).fill(0xcd);
-    const address1 = encodeBech32('alpha', 0, program1);
-    const address2 = encodeBech32('alpha', 0, program2);
-
-    const hash1 = addressToScriptHash(address1);
-    const hash2 = addressToScriptHash(address2);
-
-    expect(hash1).not.toBe(hash2);
-  });
-
-  it('should throw for invalid bech32 address', () => {
-    expect(() => addressToScriptHash('invalid')).toThrow('Invalid bech32 address');
-    expect(() => addressToScriptHash('')).toThrow('Invalid bech32 address');
-  });
-
-  it('should work with bc1 addresses (Bitcoin testnet)', () => {
-    // bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4
-    const address = 'bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4';
-    const scripthash = addressToScriptHash(address);
-
-    expect(scripthash).toHaveLength(64);
-  });
-
-  it('should produce reversed byte order (Electrum format)', () => {
-    // The scripthash should be SHA256(scriptPubKey) with reversed byte order
-    const program = new Uint8Array(20).fill(0);
-    const address = encodeBech32('alpha', 0, program);
-
-    const scripthash = addressToScriptHash(address);
-
-    // Verify it's different from non-reversed SHA256
-    // This is tested implicitly by the format requirement
-    expect(scripthash).toHaveLength(64);
-  });
-});
-
-// =============================================================================
-// ScriptPubKey Construction Tests
-// =============================================================================
-
-describe('ScriptPubKey construction', () => {
-  it('should use P2WPKH format (OP_0 + 20-byte hash)', () => {
-    // The scripthash is computed from "0014" + pubkey_hash
-    // This is the P2WPKH scriptPubKey format
-    const program = new Uint8Array([
-      0x75, 0x1e, 0x76, 0xe8, 0x19, 0x91, 0x96, 0xd4, 0x54, 0x94, 0x1c, 0x45, 0xd1, 0xb3, 0xa3,
-      0x23, 0xf1, 0x43, 0x3b, 0xd6,
-    ]);
-    const address = encodeBech32('alpha', 0, program);
-
-    // This should not throw and should produce valid scripthash
-    const scripthash = addressToScriptHash(address);
-    expect(scripthash).toHaveLength(64);
-  });
-});
